Don't attempt to evaluate non-applied lambdas
There used to be an error in this case, which would say something like "there is no function called 'lambda'", but now it works and just doesn't attempt to evaluate it. This also means two functions (non-builtin) can be equality-ed against each other. Of course, it's only syntactic equality - semantic equality would be undecideable.