Generalize builtin eval strategies
Strategies are now functions, Term -> Bool. Here's the comment for them:

"Specifies the evaluation strategy of a particular parameter of a builtin function. A strategy is used when an argument to a builtin is evaluated -- the argument will be reduced until the strategy predicate evaluates to true. In this sense, the strategy specifies some criteria which an argument must meet before being passed to it."

This allows for more advanced strategies, such as evaluating until the argument is a cons-list (like in the head builtin).