Fix a bug
When evaluating an application of a builtin, the argument would be evaluated infinitely so long as the predicate remains false and no repeats are made.

If the argument to a strict function is a builtin function, a problem arises, because they can't be checked for equality against each other and hence the application will reduce infinitely, making an infinite loop.

This was fixed by stopping the reduction loop if the term is in normal form, as well as the previous options.